refactor: consolidate utils::watcher_env_tracker to utils::env
