+++ cli/lib.rs
@@ -68,10 +68,10 @@ use crate::args::Flags;
 use crate::args::flags_from_vec_with_initial_cwd;
 use crate::args::get_default_v8_flags;
 use crate::util::display;
+use crate::util::env::WatchEnvTracker;
+use crate::util::env::load_env_variables_from_env_files;
 use crate::util::v8::get_v8_flags_from_env;
 use crate::util::v8::init_v8_flags;
-use crate::util::watch_env_tracker::WatchEnvTracker;
-use crate::util::watch_env_tracker::load_env_variables_from_env_files;
 
 #[cfg(feature = "dhat-heap")]
 #[global_allocator]
