+++ cli/util/mod.rs
@@ -17,5 +17,4 @@ pub mod temp;
 pub mod text_encoding;
 pub mod unix;
 pub mod v8;
-pub mod watch_env_tracker;
 pub mod windows;
